+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For a full diff see [`0.18.0...main`][0.18.0...main].
 
+### Changed
+
+- Show books in reading streak tooltip
+
 ## [`0.18.0`][0.18.0] - 2025-08-20
 
 For a full diff see [`0.17.0...0.18.0`][0.17.0...0.18.0].